Change field type of rate from integer to float

+++ db/migrate/20200119183735_create_currency.rb
@@ -2,7 +2,7 @@ class CreateCurrency < ActiveRecord::Migration[6.0]
   def change
     create_table :currencies do |t|
       t.string :name
-      t.integer :rate
+      t.float :rate
     end
   end
 end
